+++ OpenMarket/OpenMarket/Edit/EditViewController.swift
@@ -7,21 +7,7 @@
 
 import UIKit
 
-private enum Section {
-    case main
-}
-
-final class EditViewController: UIViewController {
-    private typealias DataSource = UICollectionViewDiffableDataSource<Section, UIImage>
-    private typealias Snapshot = NSDiffableDataSourceSnapshot<Section, UIImage>
-    
-    private let networkManager = NetworkManager<Product>()
-    
-    private var mainView: EditView?
-    private var dataSource: DataSource?
-    private var snapshot: Snapshot?
-    
-    private let imagePickerController = UIImagePickerController()
+final class EditViewController: ProductViewController {
     
     private let product: Product
     
@@ -33,12 +19,6 @@ final class EditViewController: UIViewController {
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
-    
-    override func loadView() {
-        super.loadView()
-        mainView = EditView(frame: view.bounds)
-        view = mainView
-    }
 
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -48,10 +28,6 @@ final class EditViewController: UIViewController {
         registerNotification()
     }
 
-    deinit {
-        removeNotification()
-    }
-    
     private func requestData() {
         guard let id = product.id else { return }
         
@@ -84,31 +60,16 @@ final class EditViewController: UIViewController {
         }
     }
     
-    private func configureView() {
-        configureCollectionView()
-        configureNavigationBar()
-    }
-    
-    private func configureCollectionView() {
-        mainView?.collectionView.register(ProductImageCell.self, forCellWithReuseIdentifier: ProductImageCell.identifier)
-        dataSource = makeDataSource()
-        snapshot = makeSnapsnot()
-    }
-    
-    private func configureNavigationBar() {
-        let leftBarButton = UIBarButtonItem(title: "Cancel", style: .plain, target: self, action: #selector(cancelButtonDidTapped))
-        let rightBarButton = UIBarButtonItem(title: "Done", style: .plain, target: self, action: #selector(doneButtonDidTapped))
-        
-        navigationItem.leftBarButtonItem = leftBarButton
-        navigationItem.rightBarButtonItem = rightBarButton
+    override func configureNavigationBar() {
+        super.configureNavigationBar()
         navigationItem.title = "상품수정"
     }
     
-    @objc private func cancelButtonDidTapped() {
+    @objc override func cancelButtonDidTapped() {
         navigationController?.popViewController(animated: true)
     }
     
-    @objc private func doneButtonDidTapped() {
+    @objc override func doneButtonDidTapped() {
         
         guard let data = mainView?.allData() else { return }
         
@@ -128,64 +89,9 @@ final class EditViewController: UIViewController {
         }
     }
     
-    private func registerNotification() {
-        NotificationCenter.default.addObserver(self, selector: #selector(keyboardWillShow), name: UIResponder.keyboardWillShowNotification, object: nil)
-        NotificationCenter.default.addObserver(self, selector: #selector(keyboardWillHidden), name: UIResponder.keyboardWillHideNotification, object: nil)
-    }
+    // MARK: - CollectionView DataSource
     
-    private func removeNotification() {
-        NotificationCenter.default.removeObserver(self, name: UIResponder.keyboardWillShowNotification, object: nil)
-        NotificationCenter.default.removeObserver(self, name: UIResponder.keyboardWillHideNotification, object: nil)
-    }
-    
-    @objc private func keyboardWillShow(_ sender: Notification) {
-        guard let keyboardFrame = sender.userInfo?[UIResponder.keyboardFrameEndUserInfoKey] as? NSValue else {
-            return
-        }
-        
-        let keyboardHeight = keyboardFrame.cgRectValue.height
-        mainView?.productDescriptionTextView.contentInset.bottom = keyboardHeight
-    }
-    
-    @objc private func keyboardWillHidden(_ sender: Notification) {
-        mainView?.productDescriptionTextView.contentInset.bottom = 0
-    }
-    
-    private func configurePickerController() {
-        imagePickerController.delegate = self
-        imagePickerController.allowsEditing = true
-    }
-    
-    private func albumButtonTapped() {
-        imagePickerController.sourceType = .photoLibrary
-        present(imagePickerController, animated: true)
-    }
-    
-    private func cameraButtonTapped() {
-        guard UIImagePickerController.isSourceTypeAvailable(.camera) else {
-            return
-        }
-        
-        imagePickerController.sourceType = .camera
-        present(imagePickerController, animated: true)
-    }
-}
-
-// MARK: - ImageViewController Delegate
-
-extension EditViewController: UIImagePickerControllerDelegate & UINavigationControllerDelegate {
-    func imagePickerController(_ picker: UIImagePickerController, didFinishPickingMediaWithInfo info: [UIImagePickerController.InfoKey : Any]) {
-        guard let image = info[.editedImage] as? UIImage else { return }
-        
-        insertSnapshot(images: [image])
-        picker.dismiss(animated: true)
-    }
-}
-
-// MARK: - CollectionView DataSource
-
-extension EditViewController {
-    private func makeDataSource() -> DataSource? {
+    override func makeDataSource() -> DataSource? {
         guard let mainView = mainView else { return nil }
         
         let dataSource = DataSource(collectionView: mainView.collectionView) { collectionView, indexPath, itemIdentifier in
@@ -205,38 +111,5 @@ extension EditViewController {
         }
         
         return dataSource
-    }
-    
-    private func makeSnapsnot() -> Snapshot? {
-        var snapshot = dataSource?.snapshot()
-        snapshot?.deleteAllItems()
-        snapshot?.appendSections([.main])
-        return snapshot
-    }
-    
-    private func applySnapshot(images: [UIImage]) {
-        DispatchQueue.main.async { [self] in
-            snapshot?.appendItems(images)
-            guard let snapshot = snapshot else { return }
-            
-            dataSource?.apply(snapshot, animatingDifferences: false)
-        }
-    }
-    
-    private func deleteSnapshot(images: [UIImage]) {
-        snapshot?.deleteItems(images)
-        guard let snapshot = snapshot else { return }
-        
-        dataSource?.apply(snapshot)
-    }
-    
-    private func insertSnapshot(images: [UIImage]) {
-        DispatchQueue.main.async { [self] in
-            guard let lastItem = snapshot?.itemIdentifiers.last else { return }
-            snapshot?.insertItems(images, beforeItem: lastItem)
-            guard let snapshot = snapshot else { return }
-            
-            dataSource?.apply(snapshot, animatingDifferences: false)
-        }
     }
 }
