+++ OpenMarket/OpenMarket/Edit/RegisterViewController.swift
@@ -7,28 +7,8 @@
 
 import UIKit
 
-private enum Section {
-    case main
-}
-
-final class RegisterViewController: UIViewController {
-    private typealias DataSource = UICollectionViewDiffableDataSource<Section, UIImage>
-    private typealias Snapshot = NSDiffableDataSourceSnapshot<Section, UIImage>
+final class RegisterViewController: ProductViewController {
     
-    private let networkManager = NetworkManager<Product>()
-    
-    private var mainView: EditView?
-    private var dataSource: DataSource?
-    private var snapshot: Snapshot?
-    
-    private let imagePickerController = UIImagePickerController()
-    
-    override func loadView() {
-        super.loadView()
-        mainView = EditView(frame: view.bounds)
-        view = mainView
-    }
-
     override func viewDidLoad() {
         super.viewDidLoad()
         configureView()
@@ -36,37 +16,22 @@ final class RegisterViewController: UIViewController {
         configurePickerController()
         registerNotification()
     }
-
-    deinit {
-        removeNotification()
-    }
     
-    private func configureView() {
-        configureCollectionView()
-        configureNavigationBar()
-    }
-    
-    private func configureCollectionView() {
-        mainView?.collectionView.register(ProductImageCell.self, forCellWithReuseIdentifier: ProductImageCell.identifier)
+    override func configureCollectionView() {
+        super.configureCollectionView()
         mainView?.collectionView.delegate = self
-        dataSource = makeDataSource()
-        snapshot = makeSnapsnot()
     }
     
-    private func configureNavigationBar() {
-        let leftBarButton = UIBarButtonItem(title: "Cancel", style: .plain, target: self, action: #selector(cancelButtonDidTapped))
-        let rightBarButton = UIBarButtonItem(title: "Done", style: .plain, target: self, action: #selector(doneButtonDidTapped))
-        
-        navigationItem.leftBarButtonItem = leftBarButton
-        navigationItem.rightBarButtonItem = rightBarButton
+    override func configureNavigationBar() {
+        super.configureNavigationBar()
         navigationItem.title = "상품등록"
     }
     
-    @objc private func cancelButtonDidTapped() {
+    @objc override func cancelButtonDidTapped() {
         dismiss(animated: true)
     }
     
-    @objc private func doneButtonDidTapped() {
+    @objc override func doneButtonDidTapped() {
         
         guard let sendData = multipartFormData() else { return }
         
@@ -84,48 +49,6 @@ final class RegisterViewController: UIViewController {
                 AlertDirector(viewController: self).createErrorAlert(message: "데이터를 등록하지 못했습니다.")
             }
         }
-    }
-    
-    private func registerNotification() {
-        NotificationCenter.default.addObserver(self, selector: #selector(keyboardWillShow), name: UIResponder.keyboardWillShowNotification, object: nil)
-        NotificationCenter.default.addObserver(self, selector: #selector(keyboardWillHidden), name: UIResponder.keyboardWillHideNotification, object: nil)
-    }
-    
-    private func removeNotification() {
-        NotificationCenter.default.removeObserver(self, name: UIResponder.keyboardWillShowNotification, object: nil)
-        NotificationCenter.default.removeObserver(self, name: UIResponder.keyboardWillHideNotification, object: nil)
-    }
-    
-    @objc private func keyboardWillShow(_ sender: Notification) {
-        guard let keyboardFrame = sender.userInfo?[UIResponder.keyboardFrameEndUserInfoKey] as? NSValue else {
-            return
-        }
-        
-        let keyboardHeight = keyboardFrame.cgRectValue.height
-        mainView?.productDescriptionTextView.contentInset.bottom = keyboardHeight
-    }
-    
-    @objc private func keyboardWillHidden(_ sender: Notification) {
-        mainView?.productDescriptionTextView.contentInset.bottom = 0
-    }
-    
-    private func configurePickerController() {
-        imagePickerController.delegate = self
-        imagePickerController.allowsEditing = true
-    }
-    
-    private func albumButtonTapped() {
-        imagePickerController.sourceType = .photoLibrary
-        present(imagePickerController, animated: true)
-    }
-    
-    private func cameraButtonTapped() {
-        guard UIImagePickerController.isSourceTypeAvailable(.camera) else {
-            return
-        }
-        
-        imagePickerController.sourceType = .camera
-        present(imagePickerController, animated: true)
     }
     
     private func multipartFormData() -> Data? {
@@ -159,23 +82,35 @@ final class RegisterViewController: UIViewController {
         data.appendString(boundarySuffix)
         return data
     }
+    
+    // MARK: - CollectionView DataSource
+    
+    override func makeDataSource() -> DataSource? {
+        guard let mainView = mainView else { return nil }
+        
+        let dataSource = DataSource(collectionView: mainView.collectionView) { collectionView, indexPath, itemIdentifier in
+            
+            guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: ProductImageCell.identifier, for: indexPath) as? ProductImageCell else {
+                return ProductImageCell()
+            }
+            
+            cell.removeImage = { [weak self] in
+                self?.deleteSnapshot(images: [itemIdentifier])
+            }
+            
+            cell.configure(image: itemIdentifier)
+            
+            return cell
+        }
+        
+        return dataSource
+    }
 }
 
 private extension Data {
     mutating func appendString(_ string: String) {
         guard let data = string.data(using: .utf8) else { return }
         append(data)
-    }
-}
-
-// MARK: - ImageViewController Delegate
-
-extension RegisterViewController: UIImagePickerControllerDelegate & UINavigationControllerDelegate {
-    func imagePickerController(_ picker: UIImagePickerController, didFinishPickingMediaWithInfo info: [UIImagePickerController.InfoKey : Any]) {
-        guard let image = info[.editedImage] as? UIImage else { return }
-        
-        insertSnapshot(images: [image])
-        picker.dismiss(animated: true)
     }
 }
 
@@ -199,62 +134,3 @@ extension RegisterViewController: UICollectionViewDelegate {
         }
     }
 }
-
-// MARK: - CollectionView DataSource
-
-extension RegisterViewController {
-    private func makeDataSource() -> DataSource? {
-        guard let mainView = mainView else { return nil }
-        
-        let dataSource = DataSource(collectionView: mainView.collectionView) { collectionView, indexPath, itemIdentifier in
-            
-            guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: ProductImageCell.identifier, for: indexPath) as? ProductImageCell else {
-                return ProductImageCell()
-            }
-            
-            cell.removeImage = { [weak self] in
-                self?.deleteSnapshot(images: [itemIdentifier])
-            }
-            
-            cell.configure(image: itemIdentifier)
-            
-            return cell
-        }
-        
-        return dataSource
-    }
-    
-    private func makeSnapsnot() -> Snapshot? {
-        var snapshot = dataSource?.snapshot()
-        snapshot?.deleteAllItems()
-        snapshot?.appendSections([.main])
-        return snapshot
-    }
-    
-    private func applySnapshot(images: [UIImage]) {
-        DispatchQueue.main.async { [self] in
-            snapshot?.appendItems(images)
-            guard let snapshot = snapshot else { return }
-            
-            dataSource?.apply(snapshot, animatingDifferences: false)
-        }
-    }
-    
-    private func deleteSnapshot(images: [UIImage]) {
-        snapshot?.deleteItems(images)
-        guard let snapshot = snapshot else { return }
-        
-        dataSource?.apply(snapshot)
-    }
-    
-    private func insertSnapshot(images: [UIImage]) {
-        DispatchQueue.main.async { [self] in
-            guard let lastItem = snapshot?.itemIdentifiers.last else { return }
-            snapshot?.insertItems(images, beforeItem: lastItem)
-            guard let snapshot = snapshot else { return }
-            
-            dataSource?.apply(snapshot, animatingDifferences: false)
-        }
-    }
-}
-
